requirements.txt
openai
tiktoken
prompt_toolkit
msgpack
pyyaml
json-repair
aiofiles
python-ulid
requests
flask
beautifulsoup4
discord.py
python-telegram-bot
tree-sitter
tree-sitter-python
